move catch to subdir
