+++ test/test_main.cpp
@@ -1,5 +1,5 @@
 // Copyright 2017 Jan Delgado jdelgado@gmx.net
 // common main() for all unit tests
 #define CATCH_CONFIG_MAIN
-#include "catch_amalgamated.hpp"
+#include "catch2/catch_amalgamated.hpp"
